change db to postgresql

+++ db/migrations/1681007627109-SeedUser.ts
@@ -5,16 +5,14 @@ export class SeedUser1681007627109 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
     const password = await bcrypt.hash('Hosting2!', 10);
 
-    await queryRunner.query(`
-    INSERT INTO user (username, password)
-            VALUES ('admin@test.com', '${password}');
-            `);
+    await queryRunner.query(
+      ` INSERT INTO "user" (username, password) VALUES ('admin@test.com', '${password}');`,
+    );
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`
-            DELETE FROM user
-            WHERE username = 'admin@test.com';
-        `);
+    await queryRunner.query(
+      `DELETE FROM "user" WHERE username = 'admin@test.com';`,
+    );
   }
 }
